+++ linopy/constants.py
@@ -25,6 +25,7 @@ short_LESS_EQUAL = "<"
 
 SIGNS = {EQUAL, GREATER_EQUAL, LESS_EQUAL}
 SIGNS_alternative = {long_EQUAL, short_GREATER_EQUAL, short_LESS_EQUAL}
+SIGNS_pretty = {EQUAL: "=", GREATER_EQUAL: "≥", LESS_EQUAL: "≤"}
 
 sign_replace_dict = {
     long_EQUAL: EQUAL,