+++ test/test_constraint.py
@@ -19,7 +19,7 @@ from linopy.constants import (
     short_LESS_EQUAL,
     sign_replace_dict,
 )
-from linopy.constraints import AnonymousConstraint
+from linopy.constraints import Constraint
 
 
 @pytest.fixture
@@ -73,7 +73,7 @@ def test_anonymous_constraint_from_linear_expression_le(x, y):
     expr = 10 * x + y
     con = expr <= 10
     assert isinstance(con.lhs, LinearExpression)
-    assert con.sign.item() == LESS_EQUAL
+    assert (con.sign == LESS_EQUAL).all()
     assert (con.rhs == 10).all()
 
 
@@ -81,7 +81,7 @@ def test_anonymous_constraint_from_linear_expression_ge(x, y):
     expr = 10 * x + y
     con = expr >= 10
     assert isinstance(con.lhs, LinearExpression)
-    assert con.sign.item() == GREATER_EQUAL
+    assert (con.sign == GREATER_EQUAL).all()
     assert (con.rhs == 10).all()
 
 
@@ -89,28 +89,28 @@ def test_anonymous_constraint_from_linear_expression_eq(x, y):
     expr = 10 * x + y
     con = expr == 10
     assert isinstance(con.lhs, LinearExpression)
-    assert con.sign.item() == EQUAL
+    assert (con.sign == EQUAL).all()
     assert (con.rhs == 10).all()
 
 
 def test_anonymous_constraint_from_variable_le(x):
     con = x <= 10
     assert isinstance(con.lhs, LinearExpression)
-    assert con.sign.item() == LESS_EQUAL
+    assert (con.sign == LESS_EQUAL).all()
     assert (con.rhs == 10).all()
 
 
 def test_anonymous_constraint_from_variable_ge(x):
     con = x >= 10
     assert isinstance(con.lhs, LinearExpression)
-    assert con.sign.item() == GREATER_EQUAL
+    assert (con.sign == GREATER_EQUAL).all()
     assert (con.rhs == 10).all()
 
 
 def test_anonymous_constraint_from_variable_eq(x):
     con = x == 10
     assert isinstance(con.lhs, LinearExpression)
-    assert con.sign.item() == EQUAL
+    assert (con.sign == EQUAL).all()
     assert (con.rhs == 10).all()
 
 
@@ -129,7 +129,7 @@ def test_anonymous_scalar_constraint_from_linear_expression_fail(x, y):
 def test_anonymous_constraint_sel(x, y):
     expr = 10 * x + y
     con = expr <= 10
-    assert isinstance(con.sel(first=[1, 2]), AnonymousConstraint)
+    assert isinstance(con.sel(first=[1, 2]), Constraint)
 
 
 def test_constraint_from_rule(m, x, y):
@@ -140,8 +140,8 @@ def test_constraint_from_rule(m, x, y):
             return i * x[i] >= 0
 
     coords = [x.coords["first"], y.coords["second"]]
-    con = AnonymousConstraint.from_rule(m, bound, coords)
-    assert isinstance(con, AnonymousConstraint)
+    con = Constraint.from_rule(m, bound, coords)
+    assert isinstance(con, Constraint)
     assert con.lhs.nterm == 2
     repr(con)  # test repr
 
@@ -152,8 +152,8 @@ def test_constraint_from_rule_with_none_return(m, x, y):
             return i * x[i] + y[j] >= 0
 
     coords = [x.coords["first"], y.coords["second"]]
-    con = AnonymousConstraint.from_rule(m, bound, coords)
-    assert isinstance(con, AnonymousConstraint)
+    con = Constraint.from_rule(m, bound, coords)
+    assert isinstance(con, Constraint)
     assert con.lhs.nterm == 2
     assert (con.lhs.vars.loc[0, :] == -1).all()
     assert (con.lhs.vars.loc[1, :] != -1).all()
@@ -169,11 +169,11 @@ def test_constraint_coeffs_getter(c):
 
 
 def test_constraint_sign_getter(c):
-    assert c.sign.item() == GREATER_EQUAL
+    assert (c.sign == GREATER_EQUAL).all()
 
 
 def test_constraint_rhs_getter(c):
-    assert c.rhs.item() == 0
+    assert (c.rhs == 0).all()
 
 
 def test_constraint_vars_setter(c, x):
@@ -211,12 +211,12 @@ def test_constraint_lhs_setter_invalid(c):
 
 def test_constraint_sign_setter(c):
     c.sign = EQUAL
-    assert c.sign.item() == EQUAL
+    assert (c.sign == EQUAL).all()
 
 
 def test_constraint_sign_setter_alternative(c):
     c.sign = long_EQUAL
-    assert c.sign.item() == EQUAL
+    assert (c.sign == EQUAL).all()
 
 
 def test_constraint_sign_setter_invalid(c):
@@ -227,7 +227,7 @@ def test_constraint_sign_setter_invalid(c):
 
 def test_constraint_rhs_setter(c):
     c.rhs = 2
-    assert c.rhs.item() == 2
+    assert (c.rhs == 2).all()
 
 
 def test_constraint_rhs_setter_invalid(c, x):
@@ -268,8 +268,8 @@ def test_constraint_assignment_with_args(m, x, y):
     m.add_constraints(lhs, EQUAL, 0, name="c2")
     assert m.constraints["c2"].vars.notnull().all()
     assert m.constraints["c2"].coeffs.notnull().all()
-    assert m.constraints["c2"].sign == EQUAL
-    assert m.constraints["c2"].rhs == 0
+    assert (m.constraints["c2"].sign == EQUAL).all()
+    assert (m.constraints["c2"].rhs == 0).all()
 
 
 def test_constraint_assignment_with_args_valid_sign(m, x, y):
@@ -278,8 +278,8 @@ def test_constraint_assignment_with_args_valid_sign(m, x, y):
         m.add_constraints(lhs, sign, 0, name=f"c{i}")
         assert m.constraints[f"c{i}"].vars.notnull().all()
         assert m.constraints[f"c{i}"].coeffs.notnull().all()
-        assert m.constraints[f"c{i}"].sign == sign
-        assert m.constraints[f"c{i}"].rhs == 0
+        assert (m.constraints[f"c{i}"].sign == sign).all()
+        assert (m.constraints[f"c{i}"].rhs == 0).all()
 
 
 def test_constraint_assignment_with_args_alternative_sign(m, x, y):
@@ -289,8 +289,8 @@ def test_constraint_assignment_with_args_alternative_sign(m, x, y):
         m.add_constraints(lhs, sign, 0, name=f"c{i}")
         assert m.constraints[f"c{i}"].vars.notnull().all()
         assert m.constraints[f"c{i}"].coeffs.notnull().all()
-        assert m.constraints[f"c{i}"].sign == sign_replace_dict[sign]
-        assert m.constraints[f"c{i}"].rhs == 0
+        assert (m.constraints[f"c{i}"].sign == sign_replace_dict[sign]).all()
+        assert (m.constraints[f"c{i}"].rhs == 0).all()
 
 
 def test_constraint_assignment_with_args_invalid_sign(m, x, y):