+++ test/test_variable.py
@@ -48,11 +48,17 @@ def test_variable_data(x):
 
 
 def test_wrong_variable_init(m, x):
-    with pytest.raises(ValueError):
-        linopy.Variable(x.labels.values, m)
+    # without explicit name
+    with pytest.raises(TypeError):
+        linopy.Variable(x.data, m)
 
+    # wrong data type
     with pytest.raises(ValueError):
-        linopy.Variable(x.labels, None)
+        linopy.Variable(x.labels.values, m, "")
+
+    # no model
+    with pytest.raises(ValueError):
+        linopy.Variable(x.labels, None, "")
 
 
 def test_variable_getter(x, z):
@@ -140,29 +146,29 @@ def test_variable_sum(x):
 def test_variable_where(x):
     x = x.where([True] * 4 + [False] * 6)
     assert isinstance(x, linopy.variables.Variable)
-    assert x.values[9] == x.fill_value
+    assert x.labels[9] == x.fill_value["labels"]
 
     x = x.where([True] * 4 + [False] * 6, x[0])
     assert isinstance(x, linopy.variables.Variable)
-    assert x.values[9] == x[0].label
+    assert x.labels[9] == x[0].label
 
     x = x.where([True] * 4 + [False] * 6, x.loc[0])
     assert isinstance(x, linopy.variables.Variable)
-    assert x.values[9] == x[0].label
+    assert x.labels[9] == x[0].label
 
 
 def test_variable_shift(x):
     x = x.shift(first=3)
     assert isinstance(x, linopy.variables.Variable)
-    assert x.values[0] == -1
+    assert x.labels[0] == -1
 
 
 def test_variable_bfill(x):
     x = x.where([False] * 4 + [True] * 6)
     x = x.bfill("first")
     assert isinstance(x, linopy.variables.Variable)
-    assert x.values[2] == x.values[4]
-    assert x.values[2] != x.values[5]
+    assert x.labels[2] == x.labels[4]
+    assert x.labels[2] != x.labels[5]
 
 
 def test_variable_broadcast_like(x):
@@ -174,8 +180,8 @@ def test_variable_ffill(x):
     x = x.where([True] * 4 + [False] * 6)
     x = x.ffill("first")
     assert isinstance(x, linopy.variables.Variable)
-    assert x.values[9] == x.values[3]
-    assert x.values[3] != x.values[2]
+    assert x.labels[9] == x.labels[3]
+    assert x.labels[3] != x.labels[2]
 
 
 def test_variable_fillna(x):
@@ -188,4 +194,4 @@ def test_variable_sanitize(x):
     x = x.where([True] * 4 + [False] * 6, np.nan)
     x = x.sanitize()
     assert isinstance(x, linopy.variables.Variable)
-    assert x.values[9] == -1
+    assert x.labels[9] == -1